+++ {{cookiecutter.__src_folder_name}}/infra/appservice.bicep
@@ -1,18 +1,20 @@
 {# This is the appservice version of `web.bicep` selected by post_gen_project.py #}
-param scope object
-param prefix string
+param name string
 param location string = resourceGroup().location
 param tags object = {}
-param applicationInsightsConnectionString
+param identityName string
 param pythonVersion string 
 param appCommandLine string
+param keyVaultName string
 
-{% if cookiecutter.project_host == "appservice" %}
+resource applicationInsights 'Microsoft.Insights/components@2020-02-02' existing = {
+  name: applicationInsightsName
+}
+
 module appServicePlan 'core/host/appserviceplan.bicep' = {
   name: 'serviceplan'
-  scope: scope
   params: {
-    name: '${prefix}-serviceplan'
+    name: '${name}-serviceplan'
     location: location
     tags: tags
     sku: {
@@ -21,25 +23,28 @@ module appServicePlan 'core/host/appserviceplan.bicep' = {
     reserved: true
   }
 }
-{% endif %}
 
+resource webIdentity 'Microsoft.ManagedIdentity/userAssignedIdentities@2023-01-31' = {
+  name: identityName
+  location: location
+}
 
 module web 'core/host/appservice.bicep' = {
     name: 'appservice'
-    scope: resourceGroup
     params: {
-      name: '${prefix}-web'
+      name: '${name}-web'
       location: location
       tags: union(tags, {'azd-service-name': 'web'})
+      appCommandLine: appCommandLine
       appServicePlanId: appServicePlan.outputs.id
+      keyVaultName: keyVaultName
       runtimeName: 'python'
       runtimeVersion: pythonVersion
       scmDoBuildDuringDeployment: true
       ftpsState: 'Disabled'
-      appCommandLine: appCommandLine
       managedIdentity: true
       appSettings: {
-        APPLICATIONINSIGHTS_CONNECTION_STRING: applicationInsightsConnectionString
+        APPLICATIONINSIGHTS_CONNECTION_STRING: applicationInsights.properties.ConnectionString
         RUNNING_IN_PRODUCTION: 'true'
         {% if "postgres" in cookiecutter.db_resource %}
         POSTGRES_HOST: dbserver.outputs.DOMAIN_NAME
@@ -48,11 +53,14 @@ module web 'core/host/appservice.bicep' = {
         POSTGRES_PASSWORD: '@Microsoft.KeyVault(VaultName=${keyVault.outputs.name};SecretName=DBSERVERPASSWORD)'
         {% endif %}
         {% if cookiecutter.project_backend in ("django", "flask") %}
-        SECRET_KEY: '@Microsoft.KeyVault(VaultName=${keyVault.outputs.name};SecretName=SECRETKEY)'
+        SECRET_KEY: '@Microsoft.KeyVault(VaultName=${keyVaultName};SecretName=SECRETKEY)'
         {% endif %}
         {% if 'mongodb' in cookiecutter.db_resource %}
-        AZURE_COSMOS_CONNECTION_STRING: '@Micrsoft.KeyVault(VaultName=${keyVault.outputs.name};SecretName=AZURE-COSMOS-CONNECTION-STRING)'
+        AZURE_COSMOS_CONNECTION_STRING: '@Microsoft.KeyVault(VaultName=${keyVaultName};SecretName=AZURE-COSMOS-CONNECTION-STRING)'
         {% endif %}
       }
     }
   }
+
+
+  output SERVICE_WEB_IDENTITY_PRINCIPAL_ID string = web.outputs.identityPrincipalId
