+++ {{cookiecutter.__src_folder_name}}/infra/core/security/keyvault.bicep
@@ -4,7 +4,7 @@ param tags object = {}
 
 param principalId string = ''
 
-resource keyVault 'Microsoft.KeyVault/vaults@2022-07-01' = {
+resource keyVault 'Microsoft.KeyVault/vaults@2023-02-01' = {
   name: name
   location: location
   tags: tags
